+++ packages/validation-html/src/subscribers/validation-errors-custom-attribute.ts
@@ -1,4 +1,5 @@
-import { bindable, BindingMode, customAttribute, INode } from '@aurelia/runtime-html';
+import { BindingMode } from '@aurelia/runtime';
+import { bindable, customAttribute, INode } from '@aurelia/runtime-html';
 import { IValidationController, ValidationResultsSubscriber, ValidationEvent, ValidationResultTarget } from '../validation-controller';
 import { compareDocumentPositionFlat } from './common';
 import { optional } from '@aurelia/kernel';
