+++ packages/validation-html/src/validate-binding-behavior.ts
@@ -1,20 +1,13 @@
+import { ITask } from '@aurelia/platform';
 import { DI } from '@aurelia/kernel';
+import { bindingBehavior, BindingBehaviorExpression, BindingInterceptor, BindingMediator, IBinding, LifecycleFlags, Scope } from '@aurelia/runtime';
 import {
-  bindingBehavior,
-  BindingInterceptor,
-  BindingMediator,
-  LifecycleFlags,
   PropertyBinding,
-  IBinding,
-  BindingBehaviorExpression,
-  ITask,
   IPlatform,
   ICustomElementViewModel,
 } from '@aurelia/runtime-html';
 import { PropertyRule } from '@aurelia/validation';
 import { BindingWithBehavior, IValidationController, ValidationController, BindingInfo, ValidationResultsSubscriber, ValidationEvent } from './validation-controller';
-
-import type { Scope } from '@aurelia/runtime';
 
 /**
  * Validation triggers.
