+++ packages/validation/src/rules.ts
@@ -1,4 +1,5 @@
-import { Constructable, Protocol, Metadata, Class, DI, toArray } from '@aurelia/kernel';
+import { Metadata } from '@aurelia/metadata';
+import { Constructable, Protocol, Class, DI, toArray } from '@aurelia/kernel';
 import { Interpolation, PrimitiveLiteralExpression } from '@aurelia/runtime';
 import {
   IValidateable,
